# example/cmdb/docs/swagger.yaml
basePath: /
definitions:
  ent.Alert:
    properties:
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.Project:
    properties:
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ProjectEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ProjectQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  ent.ProjectEdges:
    properties:
      role_bindings:
        description: RoleBindings holds the value of the role_bindings edge.
        items:
          $ref: '#/definitions/ent.RoleBinding'
        type: array
      services:
        description: Services holds the value of the services edge.
        items:
          $ref: '#/definitions/ent.Service'
        type: array
    type: object
  ent.RoleBinding:
    properties:
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        $ref: '#/definitions/ent.RoleBindingEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the RoleBindingQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      role:
        description: Role holds the value of the "role" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  ent.RoleBindingEdges:
    properties:
      project:
        $ref: '#/definitions/ent.Project'
        description: Project holds the value of the project edge.
      service:
        $ref: '#/definitions/ent.Service'
        description: Service holds the value of the service edge.
      user:
        $ref: '#/definitions/ent.User'
        description: User holds the value of the user edge.
    type: object
  ent.Server:
    properties:
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ServerEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ServerQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      ip:
        description: IP holds the value of the "ip" field.
        type: string
      machine_type:
        description: MachineType holds the value of the "machine_type" field.
        type: string
      platform_type:
        description: PlatformType holds the value of the "platform_type" field.
        type: string
      system_type:
        description: SystemType holds the value of the "system_type" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  ent.ServerEdges:
    properties:
      services:
        description: Services holds the value of the services edge.
        items:
          $ref: '#/definitions/ent.Service'
        type: array
    type: object
  ent.Service:
    properties:
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ServiceEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ServiceQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  ent.ServiceEdges:
    properties:
      project:
        $ref: '#/definitions/ent.Project'
        description: Project holds the value of the project edge.
      role_bindings:
        description: RoleBindings holds the value of the role_bindings edge.
        items:
          $ref: '#/definitions/ent.RoleBinding'
        type: array
      servers:
        description: Servers holds the value of the servers edge.
        items:
          $ref: '#/definitions/ent.Server'
        type: array
    type: object
  ent.User:
    properties:
      create_time:
        description: CreateTime holds the value of the "create_time" field.
        type: string
      edges:
        $ref: '#/definitions/ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      phone:
        description: Phone holds the value of the "phone" field.
        type: string
      role:
        description: Role holds the value of the "role" field.
        type: string
      update_time:
        description: UpdateTime holds the value of the "update_time" field.
        type: string
    type: object
  ent.UserEdges:
    properties:
      alerts:
        description: Alerts holds the value of the alerts edge.
        items:
          $ref: '#/definitions/ent.Alert'
        type: array
      role_bindings:
        description: RoleBindings holds the value of the role_bindings edge.
        items:
          $ref: '#/definitions/ent.RoleBinding'
        type: array
    type: object
...
  entt.GetServerListData:
    properties:
      count:
        type: integer
      result:
        items:
          $ref: '#/definitions/ent.Server'
        type: array
    type: object
...
  entt.ProjectID:
    properties:
      id:
        type: integer
    type: object
...
  entt.RestReturn:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: object
    type: object
...
  entt.ServerEdges:
    properties:
      services:
        items:
          $ref: '#/definitions/entt.ServiceID'
        type: array
    type: object
  entt.ServerNode:
    properties:
      create_time:
        format: date-time
        type: string
      edges:
        $ref: '#/definitions/entt.ServerEdges'
      id:
        type: integer
      ip:
        type: string
      machine_type:
        enum:
        - physical
        - virtual
        type: string
      platform_type:
        enum:
        - zstack
        - k8s
        - openstack
        type: string
      system_type:
        enum:
        - linux
        - windows
        type: string
      update_time:
        format: date-time
        type: string
    type: object
  entt.ServerNodeNotID:
    properties:
      create_time:
        format: date-time
        type: string
      edges:
        $ref: '#/definitions/entt.ServerEdges'
      ip:
        type: string
      machine_type:
        enum:
        - physical
        - virtual
        type: string
      platform_type:
        enum:
        - zstack
        - k8s
        - openstack
        type: string
      system_type:
        enum:
        - linux
        - windows
        type: string
      update_time:
        format: date-time
        type: string
    type: object
  entt.ServiceEdges:
    properties:
      project:
        $ref: '#/definitions/entt.ProjectID'
    type: object
  entt.ServiceID:
    properties:
      id:
        type: integer
    type: object
  entt.ServiceNode:
    properties:
      create_time:
        format: date-time
        type: string
      edges:
        $ref: '#/definitions/entt.ServiceEdges'
      id:
        type: integer
      name:
        type: string
      update_time:
        format: date-time
        type: string
    type: object
...
  router.Result:
    properties:
      code:
        type: integer
      data:
        type: object
      err:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API 文档.
  title: cmdbapi
  version: "1.0"
paths:
...
  /api/usercall/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: eq_name
        type: string
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: includes
        type: array
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/router.Result'
            - properties:
                data:
                  $ref: '#/definitions/ent.User'
              type: object
...
  /server:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entt.ServerNodeNotID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.ServerNode'
              type: object
      summary: create one server
      tags:
      - Server
  /server/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.ServerNode'
              type: object
      summary: delete one server
      tags:
      - Server
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.ServerNode'
              type: object
      summary: get one server
      tags:
      - Server
    put:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entt.ServerNodeNotID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.ServerNode'
              type: object
      summary: update one server
      tags:
      - Server
  /server/{id}/services:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: includes
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.ServiceNode'
              type: object
      summary: delete one service
      tags:
      - Service by Server
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: includes
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entt.ServiceNode'
                  type: array
              type: object
      summary: get list service
      tags:
      - Service by Server
  /servers:
    delete:
      consumes:
      - application/json
      parameters:
      - collectionFormat: multi
        in: query
        items:
          type: integer
        name: ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.ServerNode'
              type: object
      summary: delete list server
      tags:
      - Server
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: includes
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.GetServerListData'
              type: object
      summary: get list server
      tags:
      - Server
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/entt.ServerNode'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entt.ServerNodeNotID'
                  type: array
              type: object
      summary: create list server
      tags:
      - Server
    put:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/entt.ServerNode'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entt.RestReturn'
            - properties:
                data:
                  $ref: '#/definitions/entt.ServerNode'
              type: object
      summary: update list server
      tags:
      - Server
...
swagger: "2.0"